+++ tests/test_actions.py
@@ -155,11 +155,11 @@ def test_zendesk_action_mixin_action_class_subdomain(
     })
 
 
-def test_gov_notify_action_mixin_action_class(
+def test_gov_notify_email_action_mixin_action_class(
     settings, form_session, spam_control, sender
 ):
     mock_client = mock.Mock(spec_set=client.APIFormsClient)
-    action = actions.GovNotifyAction(
+    action = actions.GovNotifyEmailAction(
         client=mock_client,
         template_id='123456',
         email_address='jim@example.com',
@@ -194,12 +194,12 @@ def test_gov_notify_action_mixin_action_class(
     })
 
 
-def test_gov_notify_action_mixin_action_class_no_reply_id(
+def test_gov_notify_email_action_mixin_action_class_no_reply_id(
     settings, form_session, spam_control, sender
 ):
 
     mock_client = mock.Mock(spec_set=client.APIFormsClient)
-    action = actions.GovNotifyAction(
+    action = actions.GovNotifyEmailAction(
         client=mock_client,
         template_id='123456',
         email_address='jim@example.com',
@@ -265,3 +265,37 @@ def test_pardot_action_mixin_action_class(
             }
         }
     })
+
+
+def test_gov_notify_letter_action_mixin_action_class(
+    settings, form_session,
+):
+    mock_client = mock.Mock(spec_set=client.APIFormsClient)
+    action = actions.GovNotifyLetterAction(
+        client=mock_client,
+        template_id='123456',
+        form_url='/the/form/',
+        form_session=form_session,
+    )
+    data = {
+            'address_line_1': 'The Occupier',
+            'address_line_2': '123 High Street',
+            'postcode': 'SW14 6BF',
+            'name': 'John Smith',
+        }
+    action.save(data)
+
+    assert mock_client.submit_generic.call_count == 1
+    assert mock_client.submit_generic.call_args == mock.call({
+        'data': data,
+        'meta': {
+            'action_name': 'gov-notify-letter',
+            'template_id': '123456',
+            'form_url': '/the/form/',
+            'funnel_steps': ['one', 'two'],
+            'ingress_url': 'example.com',
+            'sender': {},
+            'spam_control': {},
+        }
+    })
+