+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [5.0.0](https://pypi.org/project/directory-forms-api-client/5.0.1/) (2019-07-09)
+[Full Changelog](https://github.com/uktrade/directory-forms-api-client/pull/21/files)
+
+**Implemented enhancements:**
+
+TT-1604 send letter gov notify
+
 ## [4.1.0](https://pypi.org/project/directory-forms-api-client/4.1.0/) (2019-06-12)
 [Full Changelog](https://github.com/uktrade/directory-forms-api-client/pull/21/files)